requirements.txt
fastapi[standard]
uvicorn[standard]>=0.27
python-multipart>=0.0.9
SQLAlchemy>=2.0
psycopg2-binary>=2.9
python-dotenv>=1.0
pydantic>=2.6
pydantic-settings>=2.2
python-jose[cryptography]>=3.3
passlib[bcrypt]>=1.7
numpy>=1.26
pillow>=10.0
opencv-python-headless>=4.8
onnxruntime>=1.17
requests>=2.31